# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel V11 Patient and Appointment Management API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get appointments in a date range",
                "operationId": "da21eb2770a2e6aee1330e384b18c5a6",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Fecha de inicio (opcional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-10"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Fecha de fin (opcional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-02-15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de citas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Register a new appointment",
                "operationId": "4731d2ccb3a3386f01c53ad0b2c4b41e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "patient_id": {
                                        "description": "ID del paciente",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "description": "ID del usuario (doctor)",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "date_appointment": {
                                        "description": "Fecha de la cita",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-20"
                                    },
                                    "time_appointment": {
                                        "description": "Hora de la cita en formato 24 horas",
                                        "type": "string",
                                        "format": "time",
                                        "example": "14:30"
                                    },
                                    "reason": {
                                        "description": "Motivo de la cita",
                                        "type": "string",
                                        "example": "Consulta general"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully registered"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/patients-doctor": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get patients assigned to a doctor",
                "operationId": "6057b090b6b4e6dbbbdc7d07359d6f14",
                "responses": {
                    "200": {
                        "description": "List of patients assigned to the doctor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "patient": {
                                                "type": "string",
                                                "example": "Juan Pérez"
                                            },
                                            "date_born": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "1990-05-15"
                                            },
                                            "gender": {
                                                "type": "string",
                                                "example": "Masculino"
                                            },
                                            "doctor": {
                                                "type": "string",
                                                "example": "Dr. Carlos López"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Solo doctores tienen acceso a esta información"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the token returned at login",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}